Source/FinalDjango/moviebook/requirements.txt
django==4.1.1
djangorestframework==3.14.0
psycopg2 == 2.9.3
Pillow==9.2.0
djangorestframework-simplejwt==5.2.1